servers/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "LangGraph agents server for ANNA (browser automation, task management, etc.)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.4.11",
    "ultralytics>=8.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20.0",
    "einops>=0.7.0",
    "timm>=0.9.0",
    "langsmith-fetch>=0.3.1",
    "motor>=3.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]

[tool.hatch.build.targets.editable]
packages = ["src/agents"]

[tool.hatch.build]
sources = ["src"]

[project.optional-dependencies]
dev = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
